make sur eindex is an int

+++ model_builder.py
@@ -302,8 +302,8 @@ if __name__ == '__main__':
     for test_point_proportion in numpy.linspace(1, 0.1, 10):
 
         X_vector, y_vector, test_X_vector, test_y_vector = train_test_split(
-            raw_X_vector[0:n_points*test_point_proportion],
-            raw_y_vector[0:n_points*test_point_proportion],
+            raw_X_vector[0:int(n_points*test_point_proportion)],
+            raw_y_vector[0:int(n_points*test_point_proportion)],
             shuffle=False, test_size=HOLDBACK_PROPORTION)
         r_squared = carbon_model_pipeline.score(X_vector, y_vector)
         r_squared_test = carbon_model_pipeline.score(
